api/requirements.txt
anyio==3.6.2
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
fastapi==0.87.0
filelock==3.8.2
google-auth==2.15.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
pbr==5.11.0
platformdirs==2.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
stevedore==4.1.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
virtualenv-clone==0.5.7
